+++ invitation-auth/build.gradle
@@ -1,0 +1,8 @@
+dependencies {
+    implementation project(':invitation-domain')
+    implementation project(':invitation-infrastructure')
+
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+}
