+++ invitation-auth/src/main/java/com/nailseong/invitation/authentication/domain/LoginSession.java
@@ -1,4 +1,4 @@
-package com.nailseong.invitation.member.domain;
+package com.nailseong.invitation.authentication.domain;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
@@ -9,12 +9,18 @@ import org.springframework.data.redis.core.RedisHash;
 public class LoginSession implements Serializable {
 
     @Id
+    private final Long memberId;
     private final String username;
     private final LocalDateTime createdAt;
 
-    public LoginSession(final String username, final LocalDateTime createdAt) {
+    public LoginSession(final Long memberId, final String username, final LocalDateTime createdAt) {
+        this.memberId = memberId;
         this.username = username;
         this.createdAt = createdAt;
+    }
+
+    public Long getMemberId() {
+        return memberId;
     }
 
     public String getUsername() {
