+++ settings.gradle
@@ -4,3 +4,5 @@ include 'invitation-ui'
 include 'invitation-domain'
 include 'invitation-infrastructure'
 include 'core'
+include 'invitation-auth'
+
